# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/authority/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理"
                ],
                "summary": "创建授权",
                "parameters": [
                    {
                        "description": "创建授权",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AddAuthorityDetailDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/authority/edit": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理"
                ],
                "summary": "编辑授权",
                "parameters": [
                    {
                        "description": "编辑授权",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateAuthorityDetailDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/authority/query_by_user_type/{user_type}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理"
                ],
                "summary": "根据用户类型查询授权",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户类型",
                        "name": "user_type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GetAuthorityDetailDTO"
                        }
                    }
                }
            }
        },
        "/api/authority/query_by_user_type_and_model": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理"
                ],
                "summary": "根据用户类型和模块查询授权",
                "parameters": [
                    {
                        "description": "根据用户类型和模块查询授权",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.GetAuthorityDetailDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GetAuthorityDetailDTO"
                        }
                    }
                }
            }
        },
        "/api/authority/set_admin/{staff_id}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理"
                ],
                "summary": "设置管理员",
                "parameters": [
                    {
                        "type": "string",
                        "description": "员工ID",
                        "name": "staff_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/authority/set_normal/{staff_id}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理"
                ],
                "summary": "设置普通用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "员工ID",
                        "name": "staff_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/candidate/accept/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "候选人管理"
                ],
                "summary": "接受候选人",
                "parameters": [
                    {
                        "type": "string",
                        "description": "候选人ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/candidate/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "候选人管理"
                ],
                "summary": "创建候选人信息",
                "parameters": [
                    {
                        "description": "候选人信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CandidateCreateDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/candidate/delete/{candidate_id}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "候选人管理"
                ],
                "summary": "删除候选人信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "候选人ID",
                        "name": "candidate_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/candidate/edit": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "候选人管理"
                ],
                "summary": "更新候选人信息",
                "parameters": [
                    {
                        "description": "更新候选人信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CandidateEditDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/candidate/query_by_name/{name}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "候选人管理"
                ],
                "summary": "根据姓名查询候选人",
                "parameters": [
                    {
                        "type": "string",
                        "description": "候选人姓名",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/api/candidate/query_by_staff_id/{staff_id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "候选人管理"
                ],
                "summary": "根据员工ID查询候选人",
                "parameters": [
                    {
                        "type": "string",
                        "description": "员工ID",
                        "name": "staff_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/api/candidate/reject/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "候选人管理"
                ],
                "summary": "拒绝候选人",
                "parameters": [
                    {
                        "type": "string",
                        "description": "候选人ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "model.AddAuthorityDetailDTO": {
            "type": "object",
            "required": [
                "authority_content",
                "model",
                "name",
                "user_type"
            ],
            "properties": {
                "authority_content": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "user_type": {
                    "type": "string"
                }
            }
        },
...
        "model.CandidateCreateDTO": {
            "type": "object",
            "properties": {
                "describe": {
                    "type": "string"
                },
                "edu_level": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "experience": {
                    "type": "string"
                },
                "job_name": {
                    "type": "string"
                },
                "major": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "staff_id": {
                    "type": "string"
                }
            }
        },
        "model.CandidateEditDTO": {
            "type": "object",
            "properties": {
                "describe": {
                    "type": "string"
                },
                "edu_level": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "evaluation": {
                    "type": "string"
                },
                "experience": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "job_name": {
                    "type": "string"
                },
                "major": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "staff_id": {
                    "type": "string"
                }
            }
        },
...
        "model.GetAuthorityDetailDTO": {
            "type": "object",
            "required": [
                "model",
                "user_type"
            ],
            "properties": {
                "model": {
                    "type": "string"
                },
                "user_type": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateAuthorityDetailDTO": {
            "type": "object",
            "required": [
                "authority_content",
                "id"
            ],
            "properties": {
                "authority_content": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        }
    }
}